First successful build for LockKeys based on .Net core!
Good things:
.Net core edition builds

Things to fix:
Many UI and UX bugs
UI elements incorrect size
Missing images and icons

WHAT THIS MEANS:
The .Net Framework edition of LockKeys is depricated and will no longer be updated.

A Beta version of LockKeys will be released featuring the first build of the .Net edition of LockKeys.